style: move filter folder
